Finalizes Ledger model and DAO

+++ Solution/Ledger-API/src/main/java/com/twou/LedgerAPI/repository/LedgerRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.twou.LedgerAPI.model.Ledger;
 
 @Repository
-public interface LedgerRepository extends JpaRepository<Ledger, Integer> {
+public interface LedgerRepository extends JpaRepository<Ledger, Long> {
 
 }
